requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
langchain==0.1.4
langchain-google-genai==0.0.6
google-generativeai==0.3.2
faiss-cpu==1.7.4
scikit-learn==1.4.0
numpy==1.26.3
python-multipart==0.0.6 